backend/pyproject.toml
[tool.poetry]
name = "multiagent-rag-system"
version = "0.1.0"
description = "A multiagent system to aid in learning programming languages and other tech"
authors = ["Jaka Krajnc <444.ous@gmail.com>"]
readme = "README.md"
packages = [
    { include = "server" },
    { include = "statemachine" },
    { include = "tools" },
]

[tool.poetry.dependencies]
python = "^3.12"
black = "^24.8.0"
pydantic-settings = "^2.6.0"

[tool.isort]
profile = "black"

[tool.poetry.group.server.dependencies]
pytest = "^8.3.2"
fastapi = "^0.111.1"
websockets = "^13.0"
sqlmodel = "^0.0.22"
psycopg = "^3.2.1"
pyjwt = "^2.9.0"
greenlet = "^3.1.1"
asyncpg = "^0.29.0"
alembic = "^1.13.3"
sqlalchemy = "^2.0.36"
anyio = "^4.6.2.post1"

[tool.poetry.group.statemachine.dependencies]
langchain = "^0.2.12"
langgraph = "^0.2.15"
langchain-openai = "^0.1.20"
langchain-community = "^0.2.11"
langchain-chroma = "^0.1.2"
langchain-postgres = "^0.0.9"
langgraph-checkpoint-postgres = "^1.0.5"
chromadb = "^0.5.5"
psycopg = "^3.2.1"
psycopg-pool = "^3.2.2"

[tool.poetry.group.tools.dependencies]
beautifulsoup4 = "^4.12.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
